PR #310 review changes
